requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# AI Services
google-genai>=0.3.0

# Image/Video processing
opencv-python>=4.8.0
requests>=2.31.0

# Audio processing
gTTS>=2.4.0
pydub>=0.25.1

# Environment
python-dotenv>=1.0.0
